♻️ refactor: separate types
